+++ component/phlare/scrape/scrape_test.go
@@ -28,8 +28,9 @@ func TestComponent(t *testing.T) {
 	arg := NewDefaultArguments()
 	arg.JobName = "test"
 	c, err := New(component.Options{
-		Logger:     util.TestLogger(t),
-		Registerer: prometheus.NewRegistry(),
+		Logger:        util.TestFlowLogger(t),
+		Registerer:    prometheus.NewRegistry(),
+		OnStateChange: func(e component.Exports) {},
 	}, arg)
 	require.NoError(t, err)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -185,7 +186,7 @@ func TestUpdateWhileScraping(t *testing.T) {
 	args.ScrapeInterval = 1 * time.Second
 
 	c, err := New(component.Options{
-		Logger:        util.TestLogger(t),
+		Logger:        util.TestFlowLogger(t),
 		Registerer:    prometheus.NewRegistry(),
 		OnStateChange: func(e component.Exports) {},
 	}, args)